W-5019073 - Reverting WK cookie cleanup back to a 15.3 state.
DIVERGENCE:
This method has been reverted back to a pre-6.1 state. As implemented here, it does not
do what it's supposed to (it doesn't do anything). The changes in the Mobile SDK to
make this method work properly in 6.1 (see
https://github.com/forcedotcom/SalesforceMobileSDK-iOS/pull/2389) are causing an issue
with auth refresh in SSO/SAML setups. The underlying functionality needs to be
revisited in both the Mobile SDK and the Salesforce App.